implement coindata service combine publisher and subscriber in viewmodel load all coins

+++ CryptoProject/Models/CoinModel.swift
@@ -54,11 +54,11 @@ struct CoinModel: Identifiable, Codable {
         let image: String
         let currentPrice: Double
         let marketCap, marketCapRank, fullyDilutedValuation: Double?
-        let totalVolume, high24H, low24H: Int?
+        let totalVolume, high24H, low24H: Double?
         let priceChange24H, priceChangePercentage24H: Double?
-        let marketCapChange24H: Int?
+        let marketCapChange24H: Double?
         let marketCapChangePercentage24H: Double?
-        let circulatingSupply, totalSupply, maxSupply, ath: Int?
+        let circulatingSupply, totalSupply, maxSupply, ath: Double?
         let athChangePercentage: Double?
         let athDate: String?
         let atl, atlChangePercentage: Double?
